LPS-56451 Move methods not specific to dom4j
