+++ util-java/src/com/liferay/util/xml/XMLFormatter.java
@@ -16,10 +16,8 @@ package com.liferay.util.xml;
 
 import com.liferay.portal.kernel.io.unsync.UnsyncByteArrayOutputStream;
 import com.liferay.portal.kernel.io.unsync.UnsyncStringReader;
-import com.liferay.portal.kernel.util.CharPool;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.security.xml.SecureXMLFactoryProviderUtil;
 
 import java.io.IOException;
@@ -38,60 +36,6 @@ import org.xml.sax.XMLReader;
  * @author Alan Zimmerman
  */
 public class XMLFormatter {
-
-	public static String fixProlog(String xml) {
-
-		// LEP-1921
-
-		if (xml != null) {
-			int pos = xml.indexOf(CharPool.LESS_THAN);
-
-			if (pos > 0) {
-				xml = xml.substring(pos);
-			}
-		}
-
-		return xml;
-	}
-
-	public static String fromCompactSafe(String xml) {
-		return StringUtil.replace(xml, "[$NEW_LINE$]", StringPool.NEW_LINE);
-	}
-
-	public static String stripInvalidChars(String xml) {
-		if (Validator.isNull(xml)) {
-			return xml;
-		}
-
-		// Strip characters that are not valid in the 1.0 XML spec
-		// http://www.w3.org/TR/2000/REC-xml-20001006#NT-Char
-
-		StringBuilder sb = new StringBuilder();
-
-		for (int i = 0; i < xml.length(); i++) {
-			char c = xml.charAt(i);
-
-			if ((c == 0x9) || (c == 0xA) || (c == 0xD) ||
-				((c >= 0x20) && (c <= 0xD7FF)) ||
-				((c >= 0xE000) && (c <= 0xFFFD)) ||
-				((c >= 0x10000) && (c <= 0x10FFFF))) {
-
-				sb.append(c);
-			}
-		}
-
-		return sb.toString();
-	}
-
-	public static String toCompactSafe(String xml) {
-		return StringUtil.replace(
-			xml,
-			new String[] {
-				StringPool.RETURN_NEW_LINE, StringPool.NEW_LINE,
-				StringPool.RETURN
-			},
-			new String[] {"[$NEW_LINE$]", "[$NEW_LINE$]", "[$NEW_LINE$]"});
-	}
 
 	public static String toString(Node node) throws IOException {
 		return toString(node, StringPool.TAB);
